+++ providers/gcs/gcs.go
@@ -22,6 +22,7 @@ import (
 	"google.golang.org/api/iterator"
 	"google.golang.org/api/option"
 	htransport "google.golang.org/api/transport/http"
+	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"gopkg.in/yaml.v2"
@@ -39,9 +40,15 @@ var DefaultConfig = Config{
 
 // Config stores the configuration for gcs bucket.
 type Config struct {
-	Bucket         string             `yaml:"bucket"`
-	ServiceAccount string             `yaml:"service_account"`
-	HTTPConfig     exthttp.HTTPConfig `yaml:"http_config"`
+	Bucket         string `yaml:"bucket"`
+	ServiceAccount string `yaml:"service_account"`
+	UseGRPC        bool   `yaml:"use_grpc"`
+	// GRPCConnPoolSize controls the size of the gRPC connection pool and should only be used
+	// when direct path is not enabled.
+	// See https://pkg.go.dev/cloud.google.com/go/storage#hdr-Experimental_gRPC_API for more details
+	// on how to enable direct path.
+	GRPCConnPoolSize int                `yaml:"grpc_conn_pool_size"`
+	HTTPConfig       exthttp.HTTPConfig `yaml:"http_config"`
 }
 
 // Bucket implements the store.Bucket and shipper.Bucket interfaces against GCS.
@@ -119,7 +126,23 @@ func NewBucketWithConfig(ctx context.Context, logger log.Logger, gc Config, comp
 	}
 	opts = append(opts, option.WithHTTPClient(httpCli))
 
-	gcsClient, err := storage.NewClient(ctx, opts...)
+	return newBucket(ctx, logger, gc, opts)
+}
+
+func newBucket(ctx context.Context, logger log.Logger, gc Config, opts []option.ClientOption) (*Bucket, error) {
+	var (
+		err       error
+		gcsClient *storage.Client
+	)
+	if gc.UseGRPC {
+		opts = append(opts,
+			option.WithGRPCDialOption(grpc.WithRecvBufferPool(grpc.NewSharedBufferPool())),
+			option.WithGRPCConnectionPool(gc.GRPCConnPoolSize),
+		)
+		gcsClient, err = storage.NewGRPCClient(ctx, opts...)
+	} else {
+		gcsClient, err = storage.NewClient(ctx, opts...)
+	}
 	if err != nil {
 		return nil, err
 	}